% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dissimilarity}
\alias{dissimilarity}
\title{dissimilarity function}
\usage{
dissimilarity(n_v, n_c, simu_type, n_simulations)
}
\arguments{
\item{n_v}{n voters}

\item{n_c}{n candidates}

\item{simu_type}{simu_type}

\item{n_simulations}{number of simulations}
}
\description{
dissimilarity function
}
