% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dissimilarity_beta_unif}
\alias{dissimilarity_beta_unif}
\title{dissimilarity function beta unif}
\usage{
dissimilarity_beta_unif(n_v, n_c, alpha, beta, n_simulations)
}
\arguments{
\item{n_v}{n voters}

\item{n_c}{n candidates}

\item{alpha}{alpha}

\item{beta}{beta}

\item{n_simulations}{number of simulations}
}
\description{
dissimilarity function beta unif
}
