% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{evolving_alpha_beta}
\alias{evolving_alpha_beta}
\title{evolving_alpha_beta}
\usage{
evolving_alpha_beta(n_simulations)
}
\arguments{
\item{n_simulations}{number of simulations}
}
\description{
evolving_alpha_beta
}
