% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{experiments}
\alias{experiments}
\title{experiments function}
\usage{
experiments(n_simulations = 2)
}
\arguments{
\item{n_simulations}{number of simulations}
}
\value{
matrix
}
\description{
experiments function
}
