% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{export_experiments_to_excel}
\alias{export_experiments_to_excel}
\title{Export experiments to excel}
\usage{
export_experiments_to_excel(df)
}
\arguments{
\item{df}{dataFrame}
}
\description{
Export experiments to excel
}
