% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{nc_nv_evolving}
\alias{nc_nv_evolving}
\title{nc_nv_evolving function}
\usage{
nc_nv_evolving(n_simulations)
}
\arguments{
\item{n_simulations}{number of simulations}
}
\description{
nc_nv_evolving function
}
