% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{similarity}
\alias{similarity}
\title{similarity_matrix function}
\usage{
similarity(n_simulations = 4)
}
\arguments{
\item{n_simulations}{number of simulations}
}
\value{
similarity_matrix
}
\description{
similarity_matrix function
}
