% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dissimilarity}
\alias{dissimilarity}
\title{dissimilarity function}
\usage{
dissimilarity(test_n_v, n_simulations = 10)
}
\arguments{
\item{test_n_v}{for experiments}

\item{n_simulations}{number of simulations}
}
\value{
similarity_matrix
}
\description{
dissimilarity function
}
