% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dissimilarity}
\alias{dissimilarity}
\title{dissimilarity function}
\usage{
dissimilarity(n_simulations = 4)
}
\arguments{
\item{n_simulations}{number of simulations}
}
\value{
similarity_matrix
}
\description{
dissimilarity function
}
